Refactor evaluator structure; no behavior changes
